Defect/de434 check giving decimal fix

+++ Gateway/MinistryPlatform.Models/DonationAndDistributionRecord.cs
@@ -4,7 +4,7 @@ namespace MinistryPlatform.Models
 {
     public class DonationAndDistributionRecord
     {
-        public int DonationAmt { get; set; }
+        public decimal DonationAmt { get; set; }
 
         public int? FeeAmt { get; set; }
